bpc_backend/requirements.txt
six~=1.16.0
Faker
pillow
markdown
python-dateutil
pytz
DateTime

django~=4.2.3
django-environ
django-filter~=23.2
django-allauth
django-resized
python-decouple
django-ckeditor
django-crispy-forms
djangorestframework
dj-rest-auth
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1

django-cors-headers
crispy-bootstrap5
django-browser-reload
drf-yasg
django-phonenumber-field[phonenumbers]